chore: removing shared config
